backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=10.4
pandas
python-dotenv
google-generativeai>=0.8.3
